fix : review 내용 적용
- max -> range 적용
- 변수명 변경
- 와일드 카드 지양
- precision변경
- id 자료형 변경

+++ haul-be/src/main/java/com/hansalchai/haul/reservation/entity/CargoOption.java
@@ -27,7 +27,7 @@ import lombok.NoArgsConstructor;
 public class CargoOption extends BaseTime {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int cargoOptionId;
+	private long cargoOptionId;
 
 	@Column(nullable = false)
 	private boolean isRefrigerated;
